add review index
